backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-multipart
python-jose[cryptography]
pydantic[email]
scikit-learn
pandas
joblib
python-dotenv
slowapi
pytest
httpx
cryptography
alembic
matplotlib
seaborn
pydantic-settings
